Implement buffering for value::{get, set, del}.
